refactor: Change Paginator model into pagination package
